# Optimized and unified bridges logic.