--- a/packages/uniforms-bridge-json-schema/src/JSONSchemaBridge.ts
+++ b/packages/uniforms-bridge-json-schema/src/JSONSchemaBridge.ts
@@ -1,11 +1,15 @@
 import invariant from 'invariant';
 import cloneDeep from 'lodash/cloneDeep';
 import get from 'lodash/get';
+import isEmpty from 'lodash/isEmpty';
 import lowerCase from 'lodash/lowerCase';
 import memoize from 'lodash/memoize';
-import omit from 'lodash/omit';
 import upperFirst from 'lodash/upperFirst';
 import { Bridge, joinName } from 'uniforms';
+
+function fieldInvariant(name: string, condition: boolean): asserts condition {
+  invariant(condition, 'Field not found in schema: "%s"', name);
+}
 
 function resolveRef(reference: string, schema: Record<string, any>) {
   invariant(
@@ -28,31 +32,36 @@ function resolveRef(reference: string, schema: Record<string, any>) {
   return resolvedReference;
 }
 
-const propMapper: Record<string, string> = {
-  maxItems: 'maxCount',
-  minItems: 'minCount',
-  maximum: 'max',
-  minimum: 'min',
-  multipleOf: 'step',
-};
-
-function distinctSchema(schema: Record<string, any>) {
-  if (schema.type === 'object') {
-    return schema;
+function resolveRefIfNeeded(
+  partial: Record<string, any>,
+  schema: Record<string, any>,
+) {
+  if (partial.$ref) {
+    partial = Object.assign({}, partial, resolveRef(partial.$ref, schema));
+    delete partial.$ref;
   }
 
-  if (schema.$ref) {
-    return { ...schema, ...resolveRef(schema.$ref, schema) };
-  }
-
-  return schema;
+  return partial;
 }
 
-function extractValue(...xs: (boolean | null | string | undefined)[]) {
-  return xs.reduce((x, y) =>
-    x === false || x === null ? '' : x !== true && x !== undefined ? x : y,
-  );
-}
+const partialNames = ['allOf', 'anyOf', 'oneOf'];
+
+const propsToRemove = [
+  'default',
+  'enum',
+  'format',
+  'isRequired',
+  'title',
+  'uniforms',
+];
+
+const propsToRename: [string, string][] = [
+  ['maxItems', 'maxCount'],
+  ['maximum', 'max'],
+  ['minItems', 'minCount'],
+  ['minimum', 'min'],
+  ['multipleOf', 'step'],
+];
 
 function pathToName(path: string) {
   path = path.startsWith('/')
@@ -65,21 +74,18 @@ function pathToName(path: string) {
   return path.slice(1);
 }
 
-function toHumanLabel(label: string) {
-  return upperFirst(lowerCase(label));
-}
-
 export default class JSONSchemaBridge extends Bridge {
-  _compiledSchema: Record<string, any> = {};
+  schema: Record<string, any>;
+  _compiledSchema: Record<string, any>;
 
   constructor(
-    public schema: Record<string, any>,
+    schema: Record<string, any>,
     public validator: (model: Record<string, any>) => any,
   ) {
     super();
 
-    this.schema = distinctSchema(schema);
-    this._compiledSchema[''] = this.schema;
+    this.schema = resolveRefIfNeeded(schema, schema);
+    this._compiledSchema = { '': this.schema };
 
     // Memoize for performance and referential equality.
     this.getField = memoize(this.getField.bind(this));
@@ -88,21 +94,23 @@ export default class JSONSchemaBridge extends Bridge {
   }
 
   getError(name: string, error: any) {
+    const details = error?.details;
+    if (!Array.isArray(details)) {
+      return null;
+    }
+
     const nameParts = joinName(null, name);
     const rootName = joinName(nameParts.slice(0, -1));
     const baseName = nameParts[nameParts.length - 1];
+    const scopedError = details.find(error => {
+      const path = pathToName(error.instancePath ?? error.dataPath);
+      return (
+        name === path ||
+        (rootName === path && baseName === error.params.missingProperty)
+      );
+    });
 
-    return (
-      // FIXME: Correct type for `error`.
-      error?.details?.find?.((detail: any) => {
-        const path = pathToName(detail.instancePath ?? detail.dataPath);
-
-        return (
-          name === path ||
-          (rootName === path && baseName === detail.params.missingProperty)
-        );
-      }) || null
-    );
+    return scopedError || null;
   }
 
   getErrorMessage(name: string, error: any) {
@@ -122,114 +130,83 @@ export default class JSONSchemaBridge extends Bridge {
   }
 
   getField(name: string) {
-    return joinName(null, name).reduce((definition, next, nextIndex, array) => {
-      const previous = joinName(array.slice(0, nextIndex));
-      const isRequired = get(
-        definition,
-        'required',
-        get(this._compiledSchema, [previous, 'required'], []),
-      ).includes(next);
-
-      const _key = joinName(previous, next);
-      const _definition = this._compiledSchema[_key] || {};
+    return joinName(null, name).reduce((definition, next, index, array) => {
+      const prevName = joinName(array.slice(0, index));
+      const nextName = joinName(prevName, next);
+      const definitionCache = (this._compiledSchema[nextName] ??= {});
+      definitionCache.isRequired = !!(
+        definition.required?.includes(next) ||
+        this._compiledSchema[prevName].required?.includes(next)
+      );
 
       if (next === '$' || next === '' + parseInt(next, 10)) {
-        invariant(
-          definition.type === 'array',
-          'Field not found in schema: "%s"',
-          name,
-        );
+        fieldInvariant(name, definition.type === 'array');
         definition = Array.isArray(definition.items)
           ? definition.items[parseInt(next, 10)]
           : definition.items;
+        fieldInvariant(name, !!definition);
       } else if (definition.type === 'object') {
-        invariant(
-          definition.properties,
-          'Field properties not found in schema: "%s"',
-          name,
-        );
+        fieldInvariant(name, !!definition.properties);
         definition = definition.properties[next];
+        fieldInvariant(name, !!definition);
       } else {
-        const [{ properties: combinedDefinition = {} } = {}] = [
-          'allOf',
-          'anyOf',
-          'oneOf',
-        ]
-          .filter(key => definition[key])
-          .map(key => {
-            // FIXME: Correct type for `definition`.
-            const localDef = (definition[key] as any[]).map(subSchema =>
-              subSchema.$ref
-                ? resolveRef(subSchema.$ref, this.schema)
-                : subSchema,
-            );
-            return localDef.find(({ properties = {} }) => properties[next]);
+        let nextFound = false;
+        partialNames.forEach(partialName => {
+          definition[partialName]?.forEach((partialElement: any) => {
+            if (!nextFound) {
+              partialElement = resolveRefIfNeeded(partialElement, this.schema);
+              if (next in partialElement.properties) {
+                definition = partialElement.properties[next];
+                nextFound = true;
+              }
+            }
           });
-
-        definition = combinedDefinition[next];
-      }
-
-      invariant(definition, 'Field not found in schema: "%s"', name);
-
-      if (definition.$ref) {
-        definition = resolveRef(definition.$ref, this.schema);
-      }
-
-      ['allOf', 'anyOf', 'oneOf'].forEach(key => {
-        if (definition[key]) {
-          // FIXME: Correct type for `definition`.
-          _definition[key] = (definition[key] as any[]).map(def =>
-            def.$ref ? resolveRef(def.$ref, this.schema) : def,
-          );
-        }
-      });
-
-      // Naive computation of combined type, properties and required
-      const combinedPartials: any[] = []
-        .concat(_definition.allOf, _definition.anyOf, _definition.oneOf)
-        .filter(Boolean);
-
-      if (combinedPartials.length) {
-        const localProperties = definition.properties
-          ? { ...definition.properties }
-          : {};
-        const localRequired = definition.required
-          ? definition.required.slice()
-          : [];
-
-        combinedPartials.forEach(({ properties, required, type }) => {
-          if (properties) {
-            Object.assign(localProperties, properties);
-          }
-          if (required) {
-            localRequired.push(...required);
-          }
-          if (type && !_definition.type) {
-            _definition.type = type;
-          }
         });
 
-        if (Object.keys(localProperties).length > 0) {
-          _definition.properties = localProperties;
-        }
-        if (localRequired.length > 0) {
-          _definition.required = localRequired;
-        }
+        fieldInvariant(name, nextFound);
       }
 
-      this._compiledSchema[_key] = Object.assign(_definition, { isRequired });
+      definition = resolveRefIfNeeded(definition, this.schema);
+
+      // Naive computation of combined type, properties and required.
+      const required = definition.required ? definition.required.slice() : [];
+      const properties = definition.properties
+        ? Object.assign({}, definition.properties)
+        : {};
+
+      partialNames.forEach(partialName => {
+        definition[partialName]?.forEach((partial: any) => {
+          partial = resolveRefIfNeeded(partial, this.schema);
+
+          if (partial.required) {
+            required.push(...partial.required);
+          }
+
+          Object.assign(properties, partial.properties);
+
+          if (!definitionCache.type && partial.type) {
+            definitionCache.type = partial.type;
+          }
+        });
+      });
+
+      if (required.length > 0) {
+        definitionCache.required = required;
+      }
+
+      if (!isEmpty(properties)) {
+        definitionCache.properties = properties;
+      }
 
       return definition;
     }, this.schema);
   }
 
-  getInitialValue(name: string, props: Record<string, any> = {}): any {
-    const { default: _default, type: _type } = this.getField(name);
+  getInitialValue(name: string, props?: Record<string, any>): any {
+    const field = this.getField(name);
     const {
-      default: defaultValue = _default !== undefined
-        ? _default
-        : get(this.schema.default, name),
-      type = _type,
+      default: defaultValue = field.default ?? get(this.schema.default, name),
+      type = field.type,
     } = this._compiledSchema[name];
 
     if (defaultValue !== undefined) {
@@ -238,8 +215,8 @@ export default class JSONSchemaBridge extends Bridge {
 
     if (type === 'array') {
       const item = this.getInitialValue(joinName(name, '0'));
-      const items = props.initialCount || 0;
-      return Array(items).fill(item);
+      const items = props?.initialCount || 0;
+      return Array.from({ length: items }, () => item);
     }
 
     if (type === 'object') {
@@ -249,52 +226,65 @@ export default class JSONSchemaBridge extends Bridge {
     return undefined;
   }
 
-  getProps(name: string, props: Record<string, any> = {}) {
-    const { uniforms, ...field } = this.getField(name);
-    const { enum: enum_, isRequired, title, ...ready } = omit(
-      { ...field, ...uniforms, ...this._compiledSchema[name] },
-      ['default', 'format', 'type'],
+  getProps(name: string, fieldProps?: Record<string, any>) {
+    const field = this.getField(name);
+    const props = Object.assign(
+      {},
+      field,
+      field.uniforms,
+      this._compiledSchema[name],
     );
 
-    if (enum_) {
-      ready.allowedValues = enum_;
-    }
+    props.label ??=
+      props.title ?? upperFirst(lowerCase(joinName(null, name).slice(-1)[0]));
+
     if (field.type === 'number') {
-      ready.decimal = true;
+      props.decimal = true;
     }
-    if (uniforms && uniforms.type !== undefined) {
-      ready.type = uniforms.type;
-    }
-    if (ready.required === undefined) {
-      ready.required = isRequired;
-    }
-    ready.label = extractValue(
-      ready.label,
-      title,
-      toHumanLabel(joinName(null, name).slice(-1)[0]),
-    );
 
-    const options = props.options || ready.options;
+    if (field.uniforms?.type !== undefined) {
+      props.type = field.uniforms.type;
+    }
+
+    if (props.required === undefined) {
+      props.required = props.isRequired;
+    }
+
+    if (props.type === field.type) {
+      delete props.type;
+    }
+
+    type OptionDict = Record<string, string>;
+    type OptionList = { label: string; value: unknown }[];
+    type Options = OptionDict | OptionList;
+    const options: Options = fieldProps?.options || props.options;
     if (options) {
-      if (!Array.isArray(options)) {
-        ready.transform = (value: any) => options[value];
-        ready.allowedValues = Object.keys(options);
+      if (Array.isArray(options)) {
+        props.allowedValues = options.map(option => option.value);
+        props.transform = (value: unknown) =>
+          options.find(option => option.value === value)!.label;
       } else {
-        ready.transform = (value: any) =>
-          options.find(option => option.value === value).label;
-        ready.allowedValues = options.map(option => option.value);
+        props.allowedValues = Object.keys(options);
+        props.transform = (value: string) => options[value];
       }
+    } else if (props.enum) {
+      props.allowedValues = props.enum;
     }
 
-    Object.keys(ready).forEach(key => {
-      if (key in propMapper) {
-        const newKey = propMapper[key];
-        ready[newKey] = ready[key];
-        delete ready[key];
+    propsToRename.forEach(([key, newKey]) => {
+      if (key in props) {
+        props[newKey] = props[key];
+        delete props[key];
       }
     });
 
-    return ready;
+    propsToRemove.forEach(key => {
+      if (key in props) {
+        delete props[key];
+      }
+    });
+
+    return props;
   }
 
   getSubfields(name = '') {