--- a/packages/uniforms-bridge-simple-schema-2/src/SimpleSchema2Bridge.ts
+++ b/packages/uniforms-bridge-simple-schema-2/src/SimpleSchema2Bridge.ts
@@ -62,7 +62,7 @@ export default class SimpleSchema2Bridge extends Bridge {
       try {
         merged.defaultValue = merged.autoValue.call({ operator: null });
       } catch (_) {
-        /* ignore it */
+        // It's fine.
       }
     }