Encouraged by the previous performance-oriented PRs (#935, #939, ad #953) as well as the one with unifying bridges logic (#956), I went one step further and done it for all the bridges. Now, this not only brings better performance but also simplifies and unifies their code. Bundles are also a little bit smaller than before.

Performance-wise, I focused on `JSONSchemaBridge`, as it has the most complex `getProps` of all the bridges. As I changed the `getField` as well, I measured `getProps` both with and without `memoize`ing `getField`. Overall, the new implementation is faster without cache than the old one with caching!
|     | with `memoize`  | without `memoize` |
| - | - | - |
| old | 523ms | 834ms |
| new | 195ms | 404ms |

<details>
<summary>Benchmark</summary>

```ts
import { JSONSchemaBridge } from 'uniforms-bridge-json-schema';

const schema = {
  $schema: 'http://json-schema.org/draft-06/schema#',
  definitions: {
    address: {
      type: 'object',
      properties: {
        city: { type: 'string', uniforms: { label: false, required: false } },
        state: {
          type: 'string',
          options: [
            { label: 'Alabama', value: 'AL' },
            { label: 'Alaska', value: 'AK' },
            { label: 'Arkansas', value: 'AR' },
          ],
        },
        street: { type: 'string' },
      },
      required: ['street', 'city', 'state'],
    },
    personalData: {
      type: 'object',
      properties: {
        firstName: { $ref: '#/definitions/firstName' },
        lastName: { $ref: '#/definitions/firstName' },
      },
      required: ['lastName'],
    },
    firstName: { type: 'string', default: 'John' },
    lastName: { type: 'string' },
    recursive: {
      type: 'object',
      properties: {
        field: { type: 'string' },
        recursive: { $ref: '#/definitions/recursive' },
      },
    },
  },
  type: 'object',
  properties: {
    age: { type: 'integer', uniforms: { component: 'span' }, default: 24 },
    billingAddress: { $ref: '#/definitions/address' },
    custom: { type: 'custom' },
    dateOfBirth: {
      type: 'string',
      format: 'date-time',
    },
    dateOfBirthTuple: {
      type: 'array',
      items: [{ type: 'integer' }, { type: 'string' }, { type: 'integer' }],
    },
    email: {
      type: 'object',
      properties: {
        work: { type: 'string' },
        other: { type: 'string' },
      },
      required: ['work'],
    },
    friends: {
      type: 'array',
      items: {
        $ref: '#/definitions/personalData',
      },
    },
    hasAJob: { type: 'boolean', title: 'Currently Employed' },
    invalid: { type: 'null' },
    personalData: { $ref: '#/definitions/personalData' },
    salary: {
      type: 'number',
      options: {
        low: 6000,
        medium: 12000,
        height: 18000,
      },
    },
    shippingAddress: {
      allOf: [
        { $ref: '#/definitions/address' },
        {
          properties: {
            type: { enum: ['residential', 'business'] },
          },
          required: ['type'],
        },
      ],
    },
    complexNames: {
      type: 'object',
      properties: {
        a: {
          type: 'array',
          items: {
            type: 'object',
            properties: {
              b: {
                type: 'object',
                properties: {
                  'c-d': {
                    type: 'array',
                    items: {
                      type: 'object',
                      properties: {
                        "f/'g": { type: 'string', pattern: '[0-9]{5}' },
                        'h/"i': { type: 'string', pattern: '[0-9]{5}' },
                        'j\'/"k': { type: 'string', pattern: '[0-9]{5}' },
                      },
                    },
                  },
                },
              },
            },
          },
        },
      },
    },
    password: { type: 'string', uniforms: { type: 'password' } },
    passwordNumeric: {
      type: 'number',
      uniforms: { type: 'password' },
      maximum: 9999,
      minimum: 1000,
      multipleOf: 3,
    },
    recursive: { $ref: '#/definitions/recursive' },
    arrayWithAllOf: {
      type: 'array',
      items: {
        type: 'object',
        allOf: [{ required: ['child'] }],
        properties: { child: { type: 'string' } },
      },
      maxItems: 3,
      minItems: 1,
    },
    nonObjectAnyOf: {
      anyOf: [
        {
          const: 'alphabetic',
          type: 'string',
        },
        {
          enum: ['top', 'middle', 'bottom'],
          type: 'string',
        },
        {
          type: 'number',
          minimum: 0,
        },
      ],
    },
    nonObjectAnyOfRequired: {
      anyOf: [
        {
          const: 'alphabetic',
          type: 'string',
        },
        {
          enum: ['top', 'middle', 'bottom'],
          type: 'string',
        },
        {
          type: 'number',
          minimum: 0,
        },
      ],
    },
    objectWithoutProperties: { type: 'object' },
  },
  required: ['dateOfBirth', 'nonObjectAnyOfRequired'],
};

const validator = jest.fn();
const bridge = new JSONSchemaBridge(schema, validator);

describe('benchmark', () => {
  it.each(Array.from({ length: 25 }, (_, x) => x + 1))('is fast %i', () => {
    expect(true).toBe(true);
    for (let index = 0; index < 1000; ++index) {
      bridge.getProps('shippingAddress.type');
      bridge.getProps('shippingAddress.type', { allowedValues: [1] });
      bridge.getProps('age');
      bridge.getProps('dateOfBirth', { label: 'Death' });
      bridge.getProps('dateOfBirth', { label: true });
      bridge.getProps('hasAJob', { label: true });
      bridge.getProps('billingAddress.city');
      bridge.getProps('dateOfBirth', { label: null });
      bridge.getProps('email.work', { placeholder: 'Email' });
      bridge.getProps('email.work', { placeholder: true });
      bridge.getProps('email.work', { placeholder: null });
      bridge.getProps('email.work', { label: null, placeholder: true });
      bridge.getProps('email.work', { label: false, placeholder: true });
      bridge.getProps('salary');
      bridge.getProps('billingAddress.state');
      bridge.getProps('billingAddress.state');
      bridge.getProps('billingAddress.state');
      bridge.getProps('billingAddress.state');
      bridge.getProps('salary');
      bridge.getProps('salary');
      bridge.getProps('salary');
      bridge.getProps('salary');
      bridge.getProps('salary', { options: { minimal: 4000, avarage: 8000 } });
      bridge.getProps('salary', { options: { minimal: 4000, avarage: 8000 } });
      bridge.getProps('salary', { options: { minimal: 4000, avarage: 8000 } });
      bridge.getProps('salary', { options: { minimal: 4000, avarage: 8000 } });
      bridge.getProps('password');
      bridge.getProps('passwordNumeric');
      bridge.getProps('personalData.firstName', { x: 1, y: 1 });
      bridge.getProps('arrayWithAllOf.0.child');
      bridge.getProps('nonObjectAnyOf');
      bridge.getProps('nonObjectAnyOfRequired');
      bridge.getProps('nonObjectAnyOf');
      bridge.getProps('arrayWithAllOf');
      bridge.getProps('arrayWithAllOf');
      bridge.getProps('passwordNumeric');
      bridge.getProps('passwordNumeric');
      bridge.getProps('passwordNumeric');
    }
  });
});
```

</details>

Of course, the sheer amount of changes also impacts the behavior a little:
* `getError` now does `Array.isArray` check, instead of duck typing `.find`.
* `getProps` now no longer duplicates the logic of resolving `label` of `useField`. However, this was never working correctly (it was different between the bridges and deviated a little from the `useField` logic).
* `JSONSchemaBridge` will **never** yield `$ref` prop (previously it did _sometimes_).
* `JSONSchemaBridge` will no longer store resolved `allOf`/`anyOf`/`oneOf` references on the `_compiledSchema` property (it's "private" anyway).